+++ service/cloudfront/errors.go
@@ -86,6 +86,27 @@ const (
 	// The maximum size of a profile for field-level encryption was exceeded.
 	ErrCodeFieldLevelEncryptionProfileSizeExceeded = "FieldLevelEncryptionProfileSizeExceeded"
 
+	// ErrCodeFunctionAlreadyExists for service response error code
+	// "FunctionAlreadyExists".
+	//
+	// A function with the same name already exists in this AWS account. To create
+	// a function, you must provide a unique name. To update an existing function,
+	// use UpdateFunction.
+	ErrCodeFunctionAlreadyExists = "FunctionAlreadyExists"
+
+	// ErrCodeFunctionInUse for service response error code
+	// "FunctionInUse".
+	//
+	// Cannot delete the function because it’s attached to one or more cache behaviors.
+	ErrCodeFunctionInUse = "FunctionInUse"
+
+	// ErrCodeFunctionSizeLimitExceeded for service response error code
+	// "FunctionSizeLimitExceeded".
+	//
+	// The function is too large. For more information, see Quotas (https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html)
+	// (formerly known as limits) in the Amazon CloudFront Developer Guide.
+	ErrCodeFunctionSizeLimitExceeded = "FunctionSizeLimitExceeded"
+
 	// ErrCodeIllegalDelete for service response error code
 	// "IllegalDelete".
 	//
@@ -137,6 +158,12 @@ const (
 	// names has been specified when not allowed or list of cookie names is missing
 	// when expected.
 	ErrCodeInvalidForwardCookies = "InvalidForwardCookies"
+
+	// ErrCodeInvalidFunctionAssociation for service response error code
+	// "InvalidFunctionAssociation".
+	//
+	// A CloudFront function association is invalid.
+	ErrCodeInvalidFunctionAssociation = "InvalidFunctionAssociation"
 
 	// ErrCodeInvalidGeoRestrictionParameter for service response error code
 	// "InvalidGeoRestrictionParameter".
@@ -304,6 +331,12 @@ const (
 	// The specified profile for field-level encryption doesn't exist.
 	ErrCodeNoSuchFieldLevelEncryptionProfile = "NoSuchFieldLevelEncryptionProfile"
 
+	// ErrCodeNoSuchFunctionExists for service response error code
+	// "NoSuchFunctionExists".
+	//
+	// The function does not exist.
+	ErrCodeNoSuchFunctionExists = "NoSuchFunctionExists"
+
 	// ErrCodeNoSuchInvalidation for service response error code
 	// "NoSuchInvalidation".
 	//
@@ -378,8 +411,7 @@ const (
 	// ErrCodePreconditionFailed for service response error code
 	// "PreconditionFailed".
 	//
-	// The precondition given in one or more of the request header fields evaluated
-	// to false.
+	// The precondition in one or more of the request fields evaluated to false.
 	ErrCodePreconditionFailed = "PreconditionFailed"
 
 	// ErrCodePublicKeyAlreadyExists for service response error code
@@ -414,6 +446,12 @@ const (
 	// or more cache behaviors.
 	ErrCodeRealtimeLogConfigInUse = "RealtimeLogConfigInUse"
 
+	// ErrCodeRealtimeLogConfigOwnerMismatch for service response error code
+	// "RealtimeLogConfigOwnerMismatch".
+	//
+	// The specified real-time log configuration belongs to a different AWS account.
+	ErrCodeRealtimeLogConfigOwnerMismatch = "RealtimeLogConfigOwnerMismatch"
+
 	// ErrCodeResourceInUse for service response error code
 	// "ResourceInUse".
 	//
@@ -433,6 +471,12 @@ const (
 	// The specified CloudFront distribution is not disabled. You must disable the
 	// distribution before you can delete it.
 	ErrCodeStreamingDistributionNotDisabled = "StreamingDistributionNotDisabled"
+
+	// ErrCodeTestFunctionFailed for service response error code
+	// "TestFunctionFailed".
+	//
+	// The CloudFront function failed.
+	ErrCodeTestFunctionFailed = "TestFunctionFailed"
 
 	// ErrCodeTooManyCacheBehaviors for service response error code
 	// "TooManyCacheBehaviors".
@@ -528,6 +572,14 @@ const (
 	// (formerly known as limits) in the Amazon CloudFront Developer Guide.
 	ErrCodeTooManyDistributionsAssociatedToOriginRequestPolicy = "TooManyDistributionsAssociatedToOriginRequestPolicy"
 
+	// ErrCodeTooManyDistributionsWithFunctionAssociations for service response error code
+	// "TooManyDistributionsWithFunctionAssociations".
+	//
+	// You have reached the maximum number of distributions that are associated
+	// with a CloudFront function. For more information, see Quotas (https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html)
+	// (formerly known as limits) in the Amazon CloudFront Developer Guide.
+	ErrCodeTooManyDistributionsWithFunctionAssociations = "TooManyDistributionsWithFunctionAssociations"
+
 	// ErrCodeTooManyDistributionsWithLambdaAssociations for service response error code
 	// "TooManyDistributionsWithLambdaAssociations".
 	//
@@ -582,6 +634,22 @@ const (
 	// The maximum number of query arg profiles for field-level encryption have
 	// been created.
 	ErrCodeTooManyFieldLevelEncryptionQueryArgProfiles = "TooManyFieldLevelEncryptionQueryArgProfiles"
+
+	// ErrCodeTooManyFunctionAssociations for service response error code
+	// "TooManyFunctionAssociations".
+	//
+	// You have reached the maximum number of CloudFront function associations for
+	// this distribution. For more information, see Quotas (https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html)
+	// (formerly known as limits) in the Amazon CloudFront Developer Guide.
+	ErrCodeTooManyFunctionAssociations = "TooManyFunctionAssociations"
+
+	// ErrCodeTooManyFunctions for service response error code
+	// "TooManyFunctions".
+	//
+	// You have reached the maximum number of CloudFront functions for this AWS
+	// account. For more information, see Quotas (https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html)
+	// (formerly known as limits) in the Amazon CloudFront Developer Guide.
+	ErrCodeTooManyFunctions = "TooManyFunctions"
 
 	// ErrCodeTooManyHeadersInCachePolicy for service response error code
 	// "TooManyHeadersInCachePolicy".
@@ -737,4 +805,10 @@ const (
 	//
 	// One or more of your trusted signers don't exist.
 	ErrCodeTrustedSignerDoesNotExist = "TrustedSignerDoesNotExist"
+
+	// ErrCodeUnsupportedOperation for service response error code
+	// "UnsupportedOperation".
+	//
+	// This operation is not supported in this region.
+	ErrCodeUnsupportedOperation = "UnsupportedOperation"
 )